requirements.txt
umap-learn==0.5.4
scipy==1.11.3
pygame==2.5.2
moderngl==5.8.2
PyGLM==2.7.1

